+++ MonoGame.Framework.Content.Pipeline/Utilities/Vector4Converter.cs
@@ -23,19 +23,19 @@ namespace MonoGame.Framework.Content.Pipeline.Utilities
     {
         Vector4 IVector4Converter<byte>.ToVector4(byte value)
         {
-            var f = (float)value / (float)byte.MaxValue;
+            var f = value / (float)byte.MaxValue;
             return new Vector4(f, 0f, 0f, 1f);
         }
 
         Vector4 IVector4Converter<short>.ToVector4(short value)
         {
-            var f = (float)value / (float)short.MaxValue;
+            var f = value / (float)short.MaxValue;
             return new Vector4(f, 0f, 0f, 1f);
         }
 
         Vector4 IVector4Converter<int>.ToVector4(int value)
         {
-            var f = (float)value / (float)int.MaxValue;
+            var f = value / (float)int.MaxValue;
             return new Vector4(f, 0f, 0f, 1f);
         }
 
@@ -56,17 +56,17 @@ namespace MonoGame.Framework.Content.Pipeline.Utilities
 
         byte IVector4Converter<byte>.FromVector4(Vector4 value)
         {
-            return (byte)(value.X * (float)byte.MaxValue);
+            return (byte)(value.X * byte.MaxValue);
         }
 
         short IVector4Converter<short>.FromVector4(Vector4 value)
         {
-            return (short)(value.X * (float)short.MaxValue);
+            return (short)(value.X * short.MaxValue);
         }
 
         int IVector4Converter<int>.FromVector4(Vector4 value)
         {
-            return (int)(value.X * (float)int.MaxValue);
+            return (int)(value.X * int.MaxValue);
         }
 
         float IVector4Converter<float>.FromVector4(Vector4 value)
